fix: reject mixed Forgejo auth flags and split tea tests
Fail closed when CLI/bootstrap flags combine tea and token-env
credentials, matching config validation. Split tea_test.go so no
single subsystem test file exceeds the 300-line AGENTS.md threshold.

Generated-By: looper 0.10.5 (runner=fixer, agent=grok-build)
